+++ src/mantispy/ds/__init__.py
@@ -2,6 +2,7 @@
 
 from mantispy.ds._blobs import blobs
 from mantispy.ds._datasets import (
+    JUMP_LITE_MODELS,
     agnp,
     amish,
     bbbc021,
@@ -9,6 +10,8 @@ from mantispy.ds._datasets import (
     jump_cells,
     jump_crispr,
     jump_export,
+    jump_lite,
+    jump_lite_targets,
     jump_plate,
     jump_target2,
     luad,
@@ -29,7 +32,10 @@ __all__ = [
     "blobs",
     "chroma",
     "jump_cells",
+    "JUMP_LITE_MODELS",
     "jump_crispr",
+    "jump_lite",
+    "jump_lite_targets",
     "jump_export",
     "jump_plate",
     "jump_target2",